pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyfreeform"
version = "0.2.0"
description = "A minimalist, art-focused Python drawing library with object-oriented design"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "PyFreeform Contributors" }
]
keywords = ["drawing", "art", "svg", "graphics", "creative-coding", "freeform", "generative-art"]
dependencies = [
    "numpy>=1.20",
    "Pillow>=9.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Artistic Software",
    "Topic :: Multimedia :: Graphics",
]

[project.urls]
Homepage = "https://github.com/pyfreeform/pyfreeform"
Documentation = "https://github.com/pyfreeform/pyfreeform#readme"
Repository = "https://github.com/pyfreeform/pyfreeform"
Issues = "https://github.com/pyfreeform/pyfreeform/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
docs = [
    "mkdocs-material>=9.0",
    "mkdocs-glightbox>=0.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pyfreeform"]

[tool.pytest.ini_options]
testpaths = ["tests"]
